requirements.txt
numpy==2.4.0
pandas==2.3.3
pyarrow==22.0.0
python-dotenv==1.2.1